fix: apply cargo clippy auto-fixes across workspace
- Remove unused imports (vault, ueba)
- Fix format! in format! args (marketplace, audit)
- Fix useless conversion (middleware, db)
- Simplify async functions (shutdown)
- Add Default impls (mail rules, signatures)
- Fix or_insert_with → or_default (chat)

10 files, 12 clippy fixes applied.

+++ crates/signapps-common/src/ueba.rs
@@ -158,7 +158,7 @@ impl AnomalyDetector {
 
         // Check for off-hours access
         for behavior in &recent_behaviors {
-            let hour = behavior.timestamp.hour() as u32;
+            let hour = behavior.timestamp.hour();
             if !baseline.typical_hours.contains(&hour) && behavior.action.to_lowercase().contains("access") {
                 return Some(Anomaly {
                     user_id,
